add toggle view-stream
