+++ pages/stream/ViewStream.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ViewStream = () => {
+	return (
+		<div className="mt-8">Here I normally see my streams</div>
+	)
+}
+
+export default ViewStream
